% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{addMetaData}
\alias{addMetaData}
\title{Add metaData}
\usage{
addMetaData(object, name)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{name}{Name of the meta data}
}
\description{
Add metaData
}
