% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_umiExperiment.R
\name{addUmiSample}
\alias{addUmiSample}
\title{Add UMI sample to an existing experiment object}
\usage{
addUmiSample(object, sampleName, sampleDir, clearData = FALSE)
}
\arguments{
\item{object}{UMIexperiment object}

\item{sampleName}{Name of new sample}

\item{sampleDir}{Directory to new sample}

\item{clearData}{Should other data in UMIexperiment be cleared}
}
\value{
A UMIexperiment object
}
\description{
Add UMI sample to an existing experiment object
}
