% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{analyzeTimeSeries}
\alias{analyzeTimeSeries}
\title{Analyze time-course data}
\usage{
analyzeTimeSeries(
  object,
  filter.name = "default",
  time.var = "time",
  use.variants = FALSE,
  group.by = NULL,
  do.plot = TRUE
)
}
\arguments{
\item{object}{UMIexperiment object containing meta data}

\item{filter.name}{Name of the filter to use.}

\item{time.var}{String. Name of thethe time variable. Default is "time"}

\item{use.variants}{Logical. Should pre computed variants be used? Default is FALSE.}

\item{group.by}{String. Variable for grouping data, e.g. replicates. Default is NULL.}

\item{do.plot}{Should plot be shown?}
}
\value{
A UMIexperiment object
}
\description{
Analyze time-course data
}
