% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{beta_binom}
\alias{beta_binom}
\title{Beta binomial model}
\usage{
beta_binom(n, size, shape1, shape2, limit.prob = 0.5, .dontuse.prob = NULL)
}
\arguments{
\item{n}{n}

\item{size}{size}

\item{shape1}{alpha}

\item{shape2}{beta}

\item{limit.prob}{0.5}

\item{.dontuse.prob}{NULL}
}
\value{
Numeric
}
\description{
Code was obtained from VGAM package function VGAM::rbetabinom.ab. The VGAM
package is available under the GPL-3 license and maintained by
Thomas Yee <t.yee at auckland.ac.nz>. Source code of the function is identical
to rbetabinom.ab, but the function name was changed to beta_binom.
}
\examples{
beta_binom(10,5, 0.5, 1)
beta_binom(10,2, 0.5, 1)

}
\references{
Yee TW (2015). Vector Generalized Linear and Additive Models: With an Implementation in R. Springer, New York, USA.
}
