% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{betaNLL}
\alias{betaNLL}
\title{Beta distribution}
\usage{
betaNLL(params, data)
}
\arguments{
\item{params}{Non-negative parameters of the Beta distribution.}

\item{data}{consensus.data table of a UMisample or UMIexperiment object.}
}
\value{
Negative log-likelihood for beta distribution.
}
\description{
Calculates the negative log likelihood for the beta distribution.
}
