% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{callVariants}
\alias{callVariants}
\title{Calculate variant p-values using permutation-based testing. A prior is fitted to model the background
error using maximum likelihood estimation of a beta distribution. The maximum likelihood estimate
of the beta distribution is then used to define the shape of a beta-binomial distribution used
to estimate variant P-Values. This can be interpreted as a probability for a variant to not have
arisen by chance}
\usage{
callVariants(object)
}
\arguments{
\item{object}{A UMierrorcorrect object.}
}
\value{
Object containing raw and FDR-adjusted P-Values
}
\description{
Calculate variant p-values using permutation-based testing. A prior is fitted to model the background
error using maximum likelihood estimation of a beta distribution. The maximum likelihood estimate
of the beta distribution is then used to define the shape of a beta-binomial distribution used
to estimate variant P-Values. This can be interpreted as a probability for a variant to not have
arisen by chance
}
\examples{
library(umiAnalyzer)
data <- UMIexperiment
data <- callVariants(data)
}
