% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{create.UMIsample}
\alias{create.UMIsample}
\title{Method for creating a UMI sample}
\usage{
create.UMIsample(sample.name, sample.dir)
}
\arguments{
\item{sample.name}{UMI sample object name}

\item{sample.dir}{Path to UMI sample}
}
\description{
Method for creating a UMI sample
}
