% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{createUmiExperiment}
\alias{createUmiExperiment}
\title{Method for creating a UMI experiment object}
\usage{
createUmiExperiment(experimentName, mainDir, sampleNames,
  importBam = FALSE)
}
\arguments{
\item{experimentName}{Name of the experiment}

\item{mainDir}{Main experiment directory}

\item{sampleNames}{List of sample names}

\item{importBam}{Logical. Should bam files be imported on creation? Default is False.}
}
\description{
Method for creating a UMI experiment object
}
\examples{
\dontrun{
library(umiAnalyzer)

main = system.file("extdata", package = "umiAnalyzer")
samples <- list.dirs(path = main, full.names = FALSE, recursive = FALSE)

exp1 <- createUMIexperiment(experimentName = "exp1", mainDir = main, sampleNames = samples)
}

}
