% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{createUMIexperiment_Debarcer}
\alias{createUMIexperiment_Debarcer}
\title{Method for creating a UMI experiment object}
\usage{
createUMIexperiment_Debarcer(experiment.name, main.dir, dir.names)
}
\arguments{
\item{experiment.name}{Name of the experiment}

\item{main.dir}{Main experiment directory}

\item{dir.names}{List of sample names}
}
\value{
A UMIexperiment object
}
\description{
Method for creating a UMI experiment object
}
