% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{createUmiSample}
\alias{createUmiSample}
\title{Method for creating a UMI sample}
\usage{
createUmiSample(sampleName, sampleDir, importBam = FALSE)
}
\arguments{
\item{sampleName}{UMI sample object name}

\item{sampleDir}{Path to UMI sample}

\item{importBam}{Logical. Should bam files be imported at object initilization? Default is False.}
}
\description{
Method for creating a UMI sample
}
