% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_functions.R
\name{createUMIsample_Debarcer}
\alias{createUMIsample_Debarcer}
\title{Method for creating a UMI sample}
\usage{
createUMIsample_Debarcer(sample.name, sample.dir, cons = "10")
}
\arguments{
\item{sample.name}{UMI sample object name}

\item{sample.dir}{Path to UMI sample}

\item{cons}{Consensus depth. Needs to be string; default is 10.}
}
\value{
A UMIsample object
}
\description{
Method for creating a UMI sample
}
