% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{download_template}
\alias{download_template}
\title{Download meta data template}
\usage{
download_template(object)
}
\arguments{
\item{object}{A UMIexperiment object}
}
\value{
A tibble containing a metadata template
}
\description{
Function for downloading a template file containing metadata.
}
\examples{
library(umiAnalyzer)

main <- system.file("extdata", package = "umiAnalyzer")

simsen <- createUmiExperiment(main)

download_template(simsen)

}
