% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{downSample}
\alias{downSample}
\title{Downsample raw reads}
\usage{
downSample(fastq)
}
\arguments{
\item{fastq}{Path to fastq containing UMIs in header}
}
\description{
Downsample raw reads
}
