% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{filterConsensusTable}
\alias{filterConsensusTable}
\title{Filter samples and amplicons from a consensus tabel}
\usage{
filterConsensusTable(consensus.data, amplicons = NULL, samples = NULL)
}
\arguments{
\item{consensus.data}{A consensus fdata table.}

\item{amplicons}{Null or list of amplicons to use.}

\item{samples}{Null or a list of samples to use.}
}
\description{
Filter samples and amplicons from a consensus tabel
}
