% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{generateAmpliconPlots}
\alias{generateAmpliconPlots}
\title{Generate Amplicon plots}
\usage{
generateAmpliconPlots(object, do.plot = TRUE)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{do.plot}{Logical. Should plots be shown.}
}
\description{
Generate Amplicon plots
}
