% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{generateVCF}
\alias{generateVCF}
\title{Generate VCF file from UMI sample or UMI experiment object}
\usage{
generateVCF(object)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}
}
\description{
Generate VCF file from UMI sample or UMI experiment object
}
