% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{getFilterdData}
\alias{getFilterdData}
\title{Method for retrieving filtered data}
\usage{
getFilterdData(object, name, save = FALSE, outDir = getwd(),
  fileName = NULL, delim = ";")
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object.}

\item{name}{String. Name of the filter.}

\item{save}{Logical, should data be saved as csv file? Default is FALSE.}

\item{outDir}{Output directory}

\item{fileName}{Filename to be used, default is the same as 'name'}

\item{delim}{Character string denoting delimiter to be used, default is ';'.}
}
\value{
A filtered consensus table, as a tibble.
}
\description{
Method for retrieving filtered data
}
