% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{getMetaData}
\alias{getMetaData}
\title{Retrieve meta data by name.}
\usage{
getMetaData(object, attributeName)
}
\arguments{
\item{object}{R object from which to get meta data.}

\item{attributeName}{Name of the meta data attribute.}
}
\description{
Retrieve meta data by name.
}
