% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{importBedFile}
\alias{importBedFile}
\title{Import bed file}
\usage{
importBedFile(path)
}
\arguments{
\item{path}{path to bed file}
}
\value{
A table containing genome positions
}
\description{
Import bed file
}
