% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{is_Numeric}
\alias{is_Numeric}
\title{Beta binomial model}
\usage{
is_Numeric(x, length.arg = Inf, integer.valued = FALSE, positive = FALSE)
}
\arguments{
\item{x}{x}

\item{length.arg}{Inf}

\item{integer.valued}{FALSE}

\item{positive}{FALSE}
}
\description{
VGAM package function VGAM:::is.Numeric
}
\references{
Yee TW (2015). Vector Generalized Linear and Additive Models: With an Implementation in R. Springer, New York, USA.
}
