% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{merge_technical_replicates}
\alias{merge_technical_replicates}
\title{A function to merge replicates in UMIexperiment object}
\usage{
merge_technical_replicates(object, filter.name)
}
\arguments{
\item{object}{UMI.experiment to which to add metadata}

\item{filter.name}{Name of the filter to use.}
}
\description{
A function to merge replicates in UMIexperiment object
}
