% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{mergeAssays}
\alias{mergeAssays}
\title{Merge assays}
\usage{
mergeAssays(object, name, assay.list)
}
\arguments{
\item{object}{A UMIexperiment object}

\item{name}{Name of the new assay}

\item{assay.list}{List of assays to merge}
}
\value{
merged consensus data
}
\description{
Merge assays together by name. Requires a name of the new assay and
a list of assays that will be merged.
}
