% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{parseBamFiles}
\alias{parseBamFiles}
\title{Function to parse bam files}
\usage{
parseBamFiles(mainDir, sampleNames, consDepth = 0)
}
\arguments{
\item{mainDir}{Directory containing umierrorcorrect output folders.}

\item{sampleNames}{A list of sample names.}

\item{consDepth}{Only retain consensus reads of at least cons.depth. Default is 0.}
}
\description{
Function to parse bam files
}
