% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{readBamFile}
\alias{readBamFile}
\title{readBamFile}
\usage{
readBamFile(sampleDir, consDepth = 0)
}
\arguments{
\item{sampleDir}{Path to UMI sample}

\item{consDepth}{Only retain consensus reads of at least cons.depth. Default is 0.}
}
\description{
Method for reading bam files using scanBam from the Rsamtools package.
}
