% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmiVisualiser.R
\name{runUmiVisualiser}
\alias{runUmiVisualiser}
\title{Function to run the umiVisualizer shiny app}
\usage{
runUmiVisualiser(docker = FALSE, path = NULL)
}
\arguments{
\item{docker}{Boolean. Are you running in a docker container?}

\item{path}{Path to directory containing UMIErrorCorrect output.}
}
\description{
Function to run the umiVisualizer shiny app
}
