% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmiVisualizer.R
\name{runUmiVisualizer}
\alias{runUmiVisualizer}
\title{Function to run the umiVisualizer shiny app}
\usage{
runUmiVisualizer(docker = FALSE, path = NULL)
}
\arguments{
\item{docker}{Boolean. Are you running in a docker container?}

\item{path}{Path to directory containing UMIErrorCorrect output.}
}
\value{
Opens the umiVisualizer app
}
\description{
Function to run the umiVisualizer shiny app
}
