% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{select_theme}
\alias{select_theme}
\title{Theme selection}
\usage{
select_theme(theme)
}
\arguments{
\item{theme}{User supplied theme selection}
}
\value{
A ggplot theme.
}
\description{
Function to select plotting theme based on user choice.
}
