% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{stacked_amplicon_plot}
\alias{stacked_amplicon_plot}
\title{Plot all variant allele bases}
\usage{
stacked_amplicon_plot(
  cons.data,
  theme = "classic",
  plot.ref = FALSE,
  abs.count = FALSE,
  option = "Pastel1",
  direction = 1
)
}
\arguments{
\item{cons.data}{Consensus data table}

\item{theme}{Plotting theme}

\item{plot.ref}{If true, shows reference base on x-axis}

\item{abs.count}{Plot absolute counts instead of frequencies.}

\item{option}{Color scheme}

\item{direction}{Direction of the color scale}
}
\value{
A ggplot object.
}
\description{
Plot all variant allele bases
}
