% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{timeSeriesGrid}
\alias{timeSeriesGrid}
\title{Plot time series data}
\usage{
timeSeriesGrid(
  object,
  filter.name = "default",
  cut.off = 5,
  min.count = 0,
  min.vaf = 0,
  amplicons = NULL,
  samples = NULL,
  x_variable = NULL,
  y_variable = "Max Non-ref Allele Frequency",
  columns = "Sample Name",
  rows = "Name",
  color_by = "Name",
  fdr = 0.05,
  do.plot = TRUE,
  use.caller = TRUE,
  bed_positions = NULL
)
}
\arguments{
\item{object}{A consensus data table}

\item{filter.name}{"default"}

\item{cut.off}{5}

\item{min.count}{0}

\item{min.vaf}{0}

\item{amplicons}{NULL}

\item{samples}{NULL}

\item{x_variable}{NULL}

\item{y_variable}{"Max Non-ref Allele Frequency"}

\item{columns}{"Sample Name"}

\item{rows}{"Name"}

\item{color_by}{"Name"}

\item{fdr}{0.05}

\item{do.plot}{Logical. Should plot be shown?}

\item{use.caller}{TRUE}

\item{bed_positions}{NULL}
}
\value{
A ggplot object.
}
\description{
Function for plotting time series or other meta data. Uses facet wrap to
display user-provided categorical variables.
}
