% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{UMIexperiment-class}
\alias{UMIexperiment-class}
\alias{UMIexperiment}
\title{UMIexperiment class}
\description{
The UMIexperiment is the core data object, storing all data and relevant
analysis data associated with your experiment. Each object has number of
slots storing raw data, graphs and processed data.
}
\section{Slots}{

\describe{
\item{\code{name}}{Optional project name for record keeping.}

\item{\code{cons.data}}{The raw consensus data supplied by the user.}

\item{\code{summary.data}}{Summary data from UMIErrorCorrect}

\item{\code{raw.error}}{Cons0 error profile}

\item{\code{reads}}{Consensus reads imported using the parseBamFiles function.}

\item{\code{meta.data}}{Sample data optionally supplied by the user.}

\item{\code{filters}}{A list of filtered cons.data, which can be accessed separately.}

\item{\code{plots}}{A list of generated plots.}

\item{\code{variants}}{Consensus table generated with the umiAnalyzer variant caller.}

\item{\code{merged.data}}{Data generated using the mergeTechnicalReplicates function.}
}}

