% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{UMIsample-class}
\alias{UMIsample-class}
\alias{UMIsample}
\title{UMIsample class}
\value{
An object of class UMIsample
}
\description{
UMIsample class
}
\section{Slots}{

\describe{
\item{\code{name}}{Sample name}

\item{\code{cons.data}}{Raw consensus data}

\item{\code{summary.data}}{Summary data from UMIErrorCorrect}

\item{\code{reads}}{Consensus reads imported from a bam file.}
}}

