% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{viewNormPlot}
\alias{viewNormPlot}
\title{View count normalization}
\usage{
viewNormPlot(object, do.plot = TRUE)
}
\arguments{
\item{object}{A UMIexperiment object containing norm plots}

\item{do.plot}{should plot be shown? If false returns a grid.arrange object}
}
\value{
A ggplot object
}
\description{
View count normalization
}
