% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{viz_Merged_data}
\alias{viz_Merged_data}
\title{Generate Merged data plots}
\usage{
viz_Merged_data(object, do.plot = TRUE)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{do.plot}{Logical. Should plots be shown.}
}
\description{
Generate Merged data plots
}
