% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{viz_stacked_counts}
\alias{viz_stacked_counts}
\title{Plot coverage before and after normalization}
\usage{
viz_stacked_counts(cons.data)
}
\arguments{
\item{cons.data}{A consensus data table}
}
\value{
A ggplot object.
}
\description{
Plot coverage before and after normalization
}
