% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{vizNormalization}
\alias{vizNormalization}
\title{Plot coverage before and after normalization}
\usage{
vizNormalization(cons.data)
}
\arguments{
\item{cons.data}{Consensus data table}
}
\value{
A list of ggplot objects.
}
\description{
Plot coverage before and after normalization
}
