% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{vizStackedCounts}
\alias{vizStackedCounts}
\title{Plot counts by nucleotide change}
\usage{
vizStackedCounts(cons.data, do.plot = TRUE)
}
\arguments{
\item{cons.data}{A consensus data table}

\item{do.plot}{Logical. Should plot be shown?}
}
\value{
A ggplot object.
}
\description{
Plot counts by nucleotide change
}
