% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{addMetaData}
\alias{addMetaData}
\title{Add metaData}
\usage{
addMetaData(object, attributeName, attributeValue)
}
\arguments{
\item{object}{R object to which meta data should be added}

\item{attributeName}{Name of the meta data attribute.}

\item{attributeValue}{Meta data to be saved.}
}
\description{
Add metaData
}
