% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addMetaData}
\alias{addMetaData}
\title{Add metaData}
\usage{
addMetaData(object, attributeName, attributeValue)
}
\arguments{
\item{object}{R object to which meta data should be added}

\item{attributeName}{Name of the meta data attribute.}

\item{attributeValue}{Meta data to be saved.}
}
\value{
A UMIexperiment object
}
\description{
Add metaData
}
\examples{
\dontrun{
library(umiAnalyzer)

main <- system.file("extdata", package = "umiAnalyzer")

simsen <- createUmiExperiment(main)

metaData <- system.file("extdata", "metadata.txt", package = "umiAnalyzer")

simsen <- addMetaData(simsen,'metaData',metaData)
}

}
