% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{analyzeTimeSeries}
\alias{analyzeTimeSeries}
\title{Analyze time-course data}
\usage{
analyzeTimeSeries(object, filter.name, time.var = "time",
  use.variants = FALSE, group.by = NULL)
}
\arguments{
\item{object}{UMI.experiment containing meta data}

\item{filter.name}{Name of the filter to use.}

\item{time.var}{String. Name of thethe time variable. Default is "time"}

\item{use.variants}{Logical. Should pre computed variants be used? Default is FALSE.}

\item{group.by}{String. Variable for grouping data, eg.g. replicates. Default is NULL.}
}
\description{
Analyze time-course data
}
