% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{beta_binom}
\alias{beta_binom}
\title{Beta binomial model}
\usage{
beta_binom(n, size, shape1, shape2, limit.prob = 0.5, .dontuse.prob = NULL)
}
\arguments{
\item{n}{n}

\item{size}{size}

\item{shape1}{alpha}

\item{shape2}{beta}

\item{limit.prob}{0.5}

\item{.dontuse.prob}{NULL}
}
\value{
Numeric
}
\description{
VGAM package function VGAM::rbetabinom.ab
}
\references{
Yee TW (2015). Vector Generalized Linear and Additive Models: With an Implementation in R. Springer, New York, USA.
}
