% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{create.UMIexperiment}
\alias{create.UMIexperiment}
\title{Method for creating a UMI experiment object}
\usage{
create.UMIexperiment(experiment.name, dir.names)
}
\description{
Method for creating a UMI experiment object
}
