% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{create.UMIexperiment}
\alias{create.UMIexperiment}
\title{Method for creating a UMI experiment object}
\usage{
create.UMIexperiment(experiment.name, main.dir, dir.names)
}
\arguments{
\item{experiment.name}{Name of the experiment}

\item{main.dir}{Main experiment directory}

\item{dir.names}{List of sample names}
}
\description{
Method for creating a UMI experiment object
}
\examples{
\dontrun{
library(umiAnalyzer)

main = system.file("extdata", package = "umiAnalyzer")
sample.names <- list.dirs(path = main, full.names = FALSE, recursive = FALSE)

exp1 <- create.UMIexperiment(experiment.name = "exp1", main.dir = main, dir.names = sample.names)
}
}
