% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{createUmiSample}
\alias{createUmiSample}
\title{createUmiSample}
\usage{
createUmiSample(sampleName, sampleDir, importBam = FALSE)
}
\arguments{
\item{sampleName}{UMI sample object name}

\item{sampleDir}{Path to UMI sample folders. Must be a folder generated by umierrorcorrect}

\item{importBam}{Logical. Should bam files be imported at object initilization? Default is False.}
}
\value{
An object of class UMIsample
}
\description{
Method for creating a UMI sample from umierrorcorrect output.
}
