% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{createUMIsample}
\alias{createUMIsample}
\title{Method for creating a UMI sample}
\usage{
createUMIsample(sample.name, sample.dir, importBam = FALSE)
}
\arguments{
\item{sample.name}{UMI sample object name}

\item{sample.dir}{Path to UMI sample}

\item{importBam}{Logical. Should bam files be imported at object initilization? Default is False.}
}
\description{
Method for creating a UMI sample
}
