% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{download_template}
\alias{download_template}
\title{Download meta data template}
\usage{
download_template(object)
}
\arguments{
\item{object}{A UmiExperiment object}
}
\value{
A tibble containing a metadata template
}
\description{
Function for downloading a template file containing metadata.
}
