% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{filterUMIobject}
\alias{filterUMIobject}
\title{Method for filtering UMIexperiment and sample objects}
\usage{
filterUMIobject(object, minDepth = 3, minCoverage = 50, minFreq = 0,
  minCount = 3)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object.}

\item{minDepth}{Consensus depth to analyze. Default is 3.}

\item{minCoverage}{Mininum coverage required for amplicons. Default is 1.}

\item{minFreq}{Minimum variant allele frequency to keep. Default is 0.}

\item{minCount}{Minimum variant allele count to keep. Default is 3.}
}
\description{
Method for filtering UMIexperiment and sample objects
}
\examples{
library(umiAnalyzer)
data <- UMIexperiment
data <- filterUMIobject(data)
}
