% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{filterVariants}
\alias{filterVariants}
\title{Filter variants based on p values or depth}
\usage{
filterVariants(object, p.adjust = 0.2, minDepth = 50)
}
\arguments{
\item{object}{A UMIexperiment object}

\item{p.adjust}{Adjusted p value (FDR)}

\item{minDepth}{Minimum consensus depth required}
}
\value{
A UMIexperiment object with filtered variants
}
\description{
Filter variants based on p values or depth
}
