% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{filterVariants}
\alias{filterVariants}
\title{Filter variants based on p values or depth}
\usage{
filterVariants(object, p.adjust = 0.2, minVarCount = 5)
}
\arguments{
\item{object}{A UMIexperiment object}

\item{p.adjust}{Numeric. Adjusted p value (FDR). Default is 0.2.}

\item{minVarCount}{Integer. Minimum variant allele count. Default is 5.}
}
\value{
A UMIexperiment object with filtered variants. Can be used to generate vcf files.
}
\description{
Filter variants based on p values or depth
}
