% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{generateAmpliconPlots}
\alias{generateAmpliconPlots}
\title{Generate Amplicon plots}
\usage{
generateAmpliconPlots(object, filter.name, do.plot = TRUE,
  amplicons = NULL)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{filter.name}{Name of the filter to be plotted.}

\item{do.plot}{Logical. Should plots be shown.}

\item{amplicons}{(Optional) character vector of amplicons to be plotted.}
}
\description{
Generate Amplicon plots
}
