% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{generateAmpliconPlots}
\alias{generateAmpliconPlots}
\title{Generate Amplicon plots}
\usage{
generateAmpliconPlots(
  object,
  filter.name = "default",
  do.plot = TRUE,
  cut.off = 5,
  amplicons = NULL,
  samples = NULL,
  abs.count = FALSE,
  y_min = 0,
  y_max = NULL,
  theme = "classic",
  option = "default",
  direction = "default",
  plot.text = TRUE,
  plot.ref = TRUE,
  stack.plot = FALSE,
  classic.plot = TRUE,
  fdr = 0.05,
  use.caller = FALSE
)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{filter.name}{Name of the filter to be plotted.}

\item{do.plot}{Logical. Should plots be shown?}

\item{cut.off}{How many variant reads are necessary to consider a variant above background? Default is 5 reads.}

\item{amplicons}{(Optional) character vector of amplicons to be plotted.}

\item{samples}{(Optional) character vector of samples to be plotted.}

\item{abs.count}{Should absolute counts be plotted instead of frequencies? Default is FALSE.}

\item{y_min}{Minimum y-axis value, default is 0}

\item{y_max}{MAximum y-axis value, default is NULL (autoscale)}

\item{theme}{Plotting theme to use, default is classic.}

\item{option}{Colour palette to use.}

\item{direction}{Orientation of the colour palette.}

\item{plot.text}{Should non-references bases be indicated above the bar?}

\item{plot.ref}{If true show reference base instead of position on x-axis.}

\item{stack.plot}{Show all variant alleles in a stacked bar plot.}

\item{classic.plot}{Show classical debarcer amplicon plot with raw error.}

\item{fdr}{False-discovery-rate cut-off for variants.}

\item{use.caller}{Should data from variant caller be used? Default is FALSE}
}
\value{
A umiExperiment object containing a ggplot object with the
amplicon plot.
}
\description{
Plots variant allele frequencies or alternate allele counts for chosen
samples and assays.
}
\examples{
\dontrun{
library(umiAnalyzer)

data <- simsen
data <- filterUmiobject(data, "myfilter")

data <- generateAmpliconPlots(simsen, "myfilter")
}
}
