% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{generateQCplots}
\alias{generateQCplots}
\title{Generate QC plots}
\usage{
generateQCplots(object, do.plot = TRUE, group.by = "assay",
  assays = NULL, samples = NULL)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{do.plot}{Logical. Should plots be shown.}

\item{group.by}{String. Which variable should be used as a factor on the x-axis. Default is assay.}

\item{assays}{(Optional) user-supllied list of assays to plot. Default is all.}

\item{samples}{(Optional) user-supllied list of samples to plot. Default is all.}
}
\description{
Generate QC plots
}
