% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{generateQCplots}
\alias{generateQCplots}
\title{Generate QC plots}
\usage{
generateQCplots(
  object,
  do.plot = TRUE,
  group.by = c("assay", "sample"),
  plotDepth = 3,
  assays = NULL,
  samples = NULL,
  theme = "classic",
  option = "viridis",
  direction = "default",
  toggle_mean = TRUE,
  center = "mean",
  line_col = "blue",
  angle = 0,
  plotly = FALSE
)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{do.plot}{Logical. Should plots be shown.}

\item{group.by}{String. Which variable should be used as a factor on the x-axis. Default is assay.}

\item{plotDepth}{Which consensus depth to plot}

\item{assays}{(Optional) user-supplied list of assays to plot. Default is all.}

\item{samples}{(Optional) user-supplied list of samples to plot. Default is all.}

\item{theme}{ggplot theme to use.}

\item{option}{Colour palette to use, etiher ggplot default or viridis colours.}

\item{direction}{If viridis colours are used, choose orientation of colour scale.}

\item{toggle_mean}{Show mean or median}

\item{center}{Choose mean or median}

\item{line_col}{Choose color for mean/median line}

\item{angle}{Angle of labels on x-axis.}

\item{plotly}{Should plotly be used for rendering?}
}
\description{
Visualise the UMI count for each selected assay and sample for a given
consensus depth. This is useful to detect differences in coverage,
especially for multiplexed assays.
}
