% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{generateVCF}
\alias{generateVCF}
\title{Generate VCF file from UMI sample or UMI experiment object}
\usage{
generateVCF(object, outDir = getwd(), outFile, printAll = FALSE)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{outDir}{String. Output directory, defaults to wokring directory.}

\item{outFile}{String. Name of the output file}

\item{printAll}{Logical. Should all or only trusted variant be printed?}
}
\description{
Generate VCF file from UMI sample or UMI experiment object
}
