% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{generateVCF}
\alias{generateVCF}
\title{Generate VCF file from UMI sample or UMI experiment object}
\usage{
generateVCF(object, outFile)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{outFile}{String. Name of the output file}
}
\description{
Generate VCF file from UMI sample or UMI experiment object
}
