% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{importBedFile}
\alias{importBedFile}
\title{Import bed file}
\usage{
importBedFile(path)
}
\arguments{
\item{path}{path to bed file}
}
\value{
A table containing genome positions
}
\description{
Import bed file
}
\examples{
library(umiAnalyzer)

bed_dir <- system.file("extdata", "simple.bed", package = "umiAnalyzer")
bed <- importBedFile(path = bed_dir)

}
