% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{importDesign}
\alias{importDesign}
\title{Import experimental design meta data such as replicates, treatments, categorical variables.}
\usage{
importDesign(object, file, delim = ",")
}
\arguments{
\item{object}{UMI.experiment to which to add metadata}

\item{file}{File containing meta data}

\item{delim}{Column separator. Default is tab.}
}
\description{
Import experimental design meta data such as replicates, treatments, categorical variables.
}
