% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotFamilyHistogram}
\alias{plotFamilyHistogram}
\title{Generate consensus depth histograms}
\usage{
plotFamilyHistogram(object, xMin = 0, xMax = 100, samples = NULL)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{xMin}{Minimum consensus family size to plot, default is 0.}

\item{xMax}{Maximum consensus family size to plot. Default is 100.}

\item{samples}{List of samples to be shown.}
}
\description{
Generate consensus depth histograms
}
