% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotFamilyHistogram}
\alias{plotFamilyHistogram}
\title{Consensus depth histograms}
\usage{
plotFamilyHistogram(object, xMin = 0, xMax = 100, samples = NULL,
  option = "viridis", direction = 1, theme = "classic")
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{xMin}{Minimum consensus family size to plot, default is 0.}

\item{xMax}{Maximum consensus family size to plot. Default is 100.}

\item{samples}{List of samples to be shown.}

\item{option}{Colour scheme to use}

\item{direction}{If using viridis colours sets the orientation of colour scale.}

\item{theme}{ggplot theme to use. Defaults to classic.}
}
\description{
Generate histograms for the frequency of barcode family depths.
}
