% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotFamilyHistogram}
\alias{plotFamilyHistogram}
\title{Generate consensus depths plots}
\usage{
plotFamilyHistogram(object)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}
}
\description{
Generate consensus depths plots
}
