% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotUmiCounts}
\alias{plotUmiCounts}
\title{Plot UMI counts}
\usage{
plotUmiCounts(object, do.plot = TRUE, amplicons = NULL,
  samples = NULL)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{do.plot}{Logical. Should plots be shown.}

\item{samples}{(Optional) user-supplied list of samples to plot. Default is all.}

\item{assays}{(Optional) user-supplied list of assays to plot. Default is all.}
}
\description{
Plot UMI counts
}
