% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmiVisualiser.R
\name{runUmiVisualiser}
\alias{runUmiVisualiser}
\title{Function to run the umiVisualiser shiny app}
\usage{
runUmiVisualiser(docker = FALSE)
}
\arguments{
\item{docker}{Boolean. Are you running in a docker container?}
}
\description{
Function to run the umiVisualiser shiny app
}
