% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmiVisualizer.R
\name{runUmiVisualizer}
\alias{runUmiVisualizer}
\title{Function to run the umiVisualizer shiny app}
\usage{
runUmiVisualizer()
}
\value{
Opens the umiVisualizer app
}
\description{
Function to run the umiVisualizer shiny app
}
\examples{
\dontrun{
library(umiAnalyzer)

runUmiVisualizer()
}

}
