% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiAnalyzer_functions.R
\name{saveConsData}
\alias{saveConsData}
\title{Save consensus data}
\usage{
saveConsData(object, outDir, save = TRUE)
}
\arguments{
\item{object}{umiExperimet object}

\item{save}{Logical. Should data be saved?}

\item{outdir}{output directory}
}
\description{
Save consensus data
}
