% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsen-data.R
\docType{data}
\name{simsen}
\alias{simsen}
\title{UMIexperiment data generated with SiMSen-Seq}
\format{
An object of class \code{"UMIexperiment"}
}
\description{
UMIexperiment data generated with SiMSen-Seq
}
\keyword{datasets}
