% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{timeSeriesGrid}
\alias{timeSeriesGrid}
\title{Plot time series data}
\usage{
timeSeriesGrid(
  object,
  filter.name = "default",
  cut.off = 5,
  min.count = 0,
  min.vaf = 0,
  amplicons = NULL,
  samples = NULL,
  x_variable = NULL,
  y_variable = "Max Non-ref Allele Frequency",
  columns = "Sample Name",
  rows = "Name",
  color_by = "Name",
  do.plot = TRUE,
  use.caller = TRUE
)
}
\arguments{
\item{object}{A consensus data table}

\item{do.plot}{Logical. Should plot be shown?}
}
\value{
A ggplot object.
}
\description{
Plot time series data
}
