% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{vizMergedData}
\alias{vizMergedData}
\title{Generate Merged data plots}
\usage{
vizMergedData(object, amplicons = NULL, do.plot = TRUE)
}
\arguments{
\item{object}{Requires a UMI sample or UMI experiment object}

\item{amplicons}{(Optional) character vector of amplicons to plot.}

\item{do.plot}{Logical. Should plots be shown.}
}
\description{
Generate Merged data plots
}
